Address CR changes
